completeDevicePowerProfile/requirements.txt
Click==7.0
azure-cosmos==4.0.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Werkzeug==0.14.1
pandas==1.0.5